requirements.txt
pdf2image
python-dotenv
streamlit
google-generativeai
